Discord Debt Bot
This is a bot made to track debt, there are commands such as ;help for showing all the commands, ;pay for paying your debt, and there is even gambling! There are likely other bots that do it but I couldn't find any myself so I went ahead and did it for my servers.